+++ src/lib.rs
@@ -148,8 +148,9 @@ pub use filter::{
     color_matrix, color_matrix_op, color_matrix_pixels, component_transfer,
     component_transfer_pixels, composite_filter, composite_filter_pixels, convolve_matrix,
     convolve_matrix_pixels, gaussian_blur, gaussian_blur_pixels, morphology, morphology_pixels,
-    ColorMatrix, ColorMatrixOp, ComponentTransfer, CompositeOp, ConvolveEdgeMode, ConvolveMatrix,
-    MorphologyOp, TransferFunc, GAUSSIAN_BLUR_BOX_THRESHOLD,
+    turbulence_filter, turbulence_filter_pixels, ColorMatrix, ColorMatrixOp, ComponentTransfer,
+    CompositeOp, ConvolveEdgeMode, ConvolveMatrix, MorphologyOp, StitchTiles, TransferFunc,
+    Turbulence, TurbulenceType, GAUSSIAN_BLUR_BOX_THRESHOLD,
 };
 pub use flatten::{flatten_arc_to_cubics, flatten_path, FlatContour};
 pub use gradient::{